Added fourth option to addAttach to allow specifying filetype.
Previously, the only filetype that could be attached with addAttach via buffer was
jpg. Everything else would be saved as a .jpg and then served into the HTML as a jpg.
Now you can specify the filetype allowing attachments to be png, svg, gif, etc.
Behaviour remains the same if this option is not specified